requirements.txt
fastapi==0.110.0
strawberry-graphql==0.219.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1 